+++ cms/contrib/frontend_v5/tests/unit/slug.test.ts
@@ -1,5 +1,5 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { addSlugHandlers } from '../../src/modules/slug';
+import { addSlugHandlers } from '../../frontend/modules/slug';
 
 /**
  * Test helpers.
